Match Sybase ctlib's handling of overlong or malformatted BCP input.
Adjust existing failure modes' boundaries and introduce the
possibility of proceeding with truncated input with or without a
message (with the latter occurring in some cases that require no
conversion).  As for messages, arrange to emit them via _ctclient_msg
rather than _csclient_msg with the same (English) wording and context
as Sybase yields, modulo formal punctuation differences and the usual
discrepancies in layer and origin numbering.

To that end:
* Add an internal BLK_CONV_STATUS enum type whose values double as
  (new) message numbers.
* Cover those numbers in _ct_get_user_api_layer_error.
* Have _cs_convert additionally take a (quite possibly NULL) pointer
  to a bulk-conversion-status value to populate as appropriate.
* Keep a running count of rows sent for reporting purposes
  (determining the column number to cite by a linear scan of the
  bindinfo's columns array).

Signed-off-by: Aaron M. Ucko <ucko@ncbi.nlm.nih.gov>